Moved root level components to new directory structure
