+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { AllPaths, Paths } from "./Routing";
 import { Redirect, Route, Switch } from "react-router";
 
-import AppBar from "./AppBar";
+import AppBar from "./components/AppBar";
 import React from "react";
 import SearchResults from "./components/SearchResults";
 import SubjectView from "./subject/components/SubjectView";
