+++ src/components/Wrapper.tsx
@@ -1,8 +1,8 @@
 import { getTheme, mergeStyleSets } from "office-ui-fabric-react";
 
 import React from "react";
-import { State } from "./Reducer";
-import SubjectComponent from "./subject/components/Subject";
+import { State } from "../Reducer";
+import SubjectComponent from "../subject/components/Subject";
 import { useSelector } from "react-redux";
 import useWindowSize from "@rehooks/window-size";
